test(e2e): dismiss BootCheckGate before auth deep-link + survive post-auth reload
Two changes that make existing specs (which don't use resetApp) actually
work end-to-end, not just up to the login screen.

1. Pre-deep-link BootCheckGate dismissal in triggerAuthDeepLinkBypass.
   The handler calls `waitForAuthReadiness` which can't make progress
   while the "Choose core mode" modal is up — it eventually fails with
   "Sign-in failed. Please try again." and the spec is wedged on the
   login screen. Inlined here (rather than imported from shared-flows)
   to avoid a circular dependency.

2. walkOnboarding tolerates the post-auth reload. The deep-link auth
   triggers an identity flip → `restartApp` → reload, which re-mounts
   onboarding fresh for the new user namespace. Previously the walker
   exited the moment the next-button briefly unmounted between steps
   and the spec navigated mid-reload. Now it stays in the loop while
   the hash is still on `#/onboarding/*`.

Also factor `dismissBootCheckGateIfVisible` into shared-flows as the
canonical export and have resetApp + walkOnboarding delegate to it.

Settings - Developer Options now 2/3 passing (was 0/3); cron-jobs-flow
still 4/4. Remaining flake on the first test of a fresh launch is a
warmup race, not a structural issue.
